refactor: wines/[id] 경로로 가도록 경로 수정
